Update and rename flowers.py to casino.py
